Add guard in viewDidAppear
Add guard to only call tabbyButtonDidPress in viewDidAppear if it has
no child view controllers.

I believe this method is invoked to programmatically tap the current
tab bar to setup the view controller at that index. However, we don't
want to do this every time the tabby controller appears on screen. To
avoid this, we only call that method if there are no child view
controllers in TabbyController.